chore: extract FakeTimers into a separate package (#7987)
