+++ packages/jest-environment-jsdom/package.json
@@ -9,6 +9,7 @@
   "license": "MIT",
   "main": "build/index.js",
   "dependencies": {
+    "@jest/fake-timers": "^24.1.0",
     "jest-mock": "^24.0.0",
     "jest-util": "^24.0.0",
     "jsdom": "^11.5.1"
