+++ packages/jest-environment/tsconfig.json
@@ -5,6 +5,7 @@
     "outDir": "build"
   },
   "references": [
+    {"path": "../jest-fake-timers"},
     {"path": "../jest-transform"},
     {"path": "../jest-types"},
     {"path": "../jest-util"}
