+++ packages/jest-fake-timers/.npmignore
@@ -1,0 +1,3 @@
+**/__mocks__/**
+**/__tests__/**
+src
