+++ packages/jest-util/package.json
@@ -10,14 +10,13 @@
   "main": "build/index.js",
   "types": "build/index.d.ts",
   "dependencies": {
+    "@jest/fake-timers": "^24.1.0",
     "@jest/types": "^24.1.0",
     "@types/node": "*",
     "callsites": "^3.0.0",
     "chalk": "^2.0.1",
     "graceful-fs": "^4.1.15",
     "is-ci": "^2.0.0",
-    "jest-message-util": "^24.0.0",
-    "jest-mock": "^24.0.0",
     "mkdirp": "^0.5.1",
     "readable-stream": "^3.1.1",
     "slash": "^2.0.0",
